Move all ES client-related code into its own package.
This commit reorganizes the code formerly contained in client.py and
async_connection.py, as well as the corresponding tests. This will make
future customizations of the sync and async ES clients more readable as they
will now live in their own modules. This new client package can easily expand
in the future to include any utility functions that need to be shared between
them.
